PUBDEV-8470 - separate test for in-training checkpoints to specific file
